Direct debit progress
extracted jsx to dedicated component
